+++ confluence/content.go
@@ -45,7 +45,7 @@ type Version struct {
 
 func (c *Client) CreateContent(content *Content) (*Content, error) {
 	var response Content
-	if err := c.Post("/wiki/rest/api/content", content, &response); err != nil {
+	if err := c.Post("/rest/api/content", content, &response); err != nil {
 		return nil, err
 	}
 	return &response, nil
@@ -53,7 +53,7 @@ func (c *Client) CreateContent(content *Content) (*Content, error) {
 
 func (c *Client) GetContent(id string) (*Content, error) {
 	var response Content
-	path := fmt.Sprintf("/wiki/rest/api/content/%s?expand=space,body.storage,version,ancestors", id)
+	path := fmt.Sprintf("/rest/api/content/%s?expand=space,body.storage,version,ancestors", id)
 	if err := c.Get(path, &response); err != nil {
 		return nil, err
 	}
@@ -63,7 +63,7 @@ func (c *Client) GetContent(id string) (*Content, error) {
 func (c *Client) UpdateContent(content *Content) (*Content, error) {
 	var response Content
 	content.Version.Number++
-	path := fmt.Sprintf("/wiki/rest/api/content/%s", content.Id)
+	path := fmt.Sprintf("/rest/api/content/%s", content.Id)
 	if err := c.Put(path, content, &response); err != nil {
 		return nil, err
 	}
@@ -71,7 +71,7 @@ func (c *Client) UpdateContent(content *Content) (*Content, error) {
 }
 
 func (c *Client) DeleteContent(id string) error {
-	path := fmt.Sprintf("/wiki/rest/api/content/%s", id)
+	path := fmt.Sprintf("/rest/api/content/%s", id)
 	if err := c.Delete(path); err != nil {
 		return err
 	}