+++ confluence/resource_content.go
@@ -137,7 +137,9 @@ func updateResourceDataFromContent(d *schema.ResourceData, content *Content, cli
 		"title":   content.Title,
 		"version": content.Version.Number,
 		"url":     client.URL(content.Links.Context + content.Links.WebUI),
-		"parent":  content.Ancestors[len(content.Ancestors)-1].Id, // the last ancestor is the parent
+	}
+	if len(content.Ancestors) > 1 {
+		m["parent"] = content.Ancestors[len(content.Ancestors)-1].Id
 	}
 	for k, v := range m {
 		err := d.Set(k, v)