+++ CHANGELOG.md
@@ -1,5 +1,13 @@
-## To be released
+## 0.2.0
 
 FEATURES:
 
-* **New Resource:** `confluence_content`
+- **Experimental:** Support for Confluence Server
+- Added additional parameters for public and api sites and schemes for Confluence Server
+- Fix: Fixed index error when creating resources without a parent
+
+## 0.1.0
+
+FEATURES:
+
+- **New Resource:** `confluence_content`